+++ src/Contracts/EasyClientContract.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace EasyHttp\Contracts\Contracts;
+namespace EasyHTTP\Contracts\Contracts;
 
 interface EasyClientContract
 {
-    public function getRequest(): HttpClientRequest;
-    public function call(string $method, string $uri): HttpClientResponse;
-    public function prepareRequest(string $method, string $uri): HttpClientRequest;
+    public function getRequest(): HTTPClientRequest;
+    public function call(string $method, string $uri): HTTPClientResponse;
+    public function prepareRequest(string $method, string $uri): HTTPClientRequest;
     public function withHandler(callable $handler): self;
-    public function execute(): HttpClientResponse;
+    public function execute(): HTTPClientResponse;
 }