+++ src/Exceptions/HttpConnectionException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace EasyHttp\Contracts\Exceptions;
-
-class HttpConnectionException extends HttpClientException
-{
-}